Update Download.py : Support for DAC added

+++ Download.py
@@ -16,6 +16,13 @@ class Download :
         if self.OwnerCheck(Owner, FileName) == -1 :
            return "Permission Denied !!!\n"
 
+        file = open("Files/" + FileName, "r")
+        file_acl = file.readline()
+        file_acl = file.readline()
+        file.close()
+        if self._CheckDiscretionaryAccess(file_acl, Owner) == -1:
+            return "Permission Denied!(By discretionary access control rules)\n"
+
         os.remove("Files/" + FileName)
 
         return FileName + ".txt Removed from Server Successfully !!!\n"
@@ -39,3 +46,13 @@ class Download :
                 if set[0] == Owner :
                     return 1
                 return -1
+
+    def _CheckDiscretionaryAccess(self, acl, username):
+        index = acl.find(username + ':')
+        if index == -1:
+            return 1
+
+        user_acl = acl[ index + len(username) + 1 : index + len(username) + 4]
+        if 'g' in user_acl:
+            return 1
+        return -1
